Refactor to allow testability
* Add tests for alias flag present and absent.  All tests pass.
* Added fakehexoFactory.js to create fake hexo objects.
* Separate index.js and migrator.js, refactor migrator.js to
  export a function that can be called multiple times.

  Note: Directly executed require blocks are cached by node, so can
  only be called once. This is an issue for testing, where you
  want to rerun setup code before each test.
