+++ package.json
@@ -32,5 +32,6 @@
   },
   "scripts": {
     "test": "mocha test"
+    
   }
 }
